+++ src/QNEthernetUDP.h
@@ -70,6 +70,9 @@ class EthernetUDP final : public UDP {
   int peek() override;
   void flush() override;
 
+  // Returns a pointer to the received packet data.
+  const unsigned char *data();
+
   IPAddress remoteIP() override;
   uint16_t remotePort() override;